requirements.txt
# Core dependencies
PyYAML>=6.0.1

# Development dependencies (optional, for testing)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.0.0
